+++ packages/api/src/menus/router.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 import { createTRPCRouter, publicProcedure } from "../trpc";
-import { getMenuProcedure } from "./procedures/get";
+import { getMenuProcedure } from "./procedures/getMenu";
 
 // import { createTRPCRouter, publicProcedure } from "../trpc";
 // import { parseCampusDish } from "./parse";
